requirements.txt
# BioASQ 13b Pipeline Requirements
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
requests>=2.31.0
tqdm>=4.66.0
openai>=1.30.0
numpy>=1.26.0
